splitting FlowProblemProperties.hpp
so that more flexibility on what compositial and blackoil can do.
The main aim is to reduce code duplication.
