+++ opm/simulators/flow/FlowBaseProblemProperties.hpp
@@ -23,13 +23,12 @@
 /*!
  * \file
  *
- * \copydoc Opm::FlowProblem
+ * \copydoc Opm::FlowBaseProblem
  */
-#ifndef OPM_FLOW_PROBLEM_PROPERTIES_HPP
-#define OPM_FLOW_PROBLEM_PROPERTIES_HPP
+#ifndef OPM_FLOW_BASE_PROBLEM_PROPERTIES_HPP
+#define OPM_FLOW_BASE_PROBLEM_PROPERTIES_HPP
 
 
-#include <opm/material/fluidmatrixinteractions/EclMaterialLawManager.hpp>
 #include <opm/material/thermal/EclThermalLawManager.hpp>
 
 #include <opm/models/discretization/ecfv/ecfvdiscretization.hh>
@@ -40,10 +39,6 @@
 #include <opm/simulators/flow/DummyGradientCalculator.hpp>
 #include <opm/simulators/flow/EclWriter.hpp>
 #include <opm/simulators/flow/FlowProblemParameters.hpp>
-#include <opm/simulators/flow/FIBlackoilModel.hpp>
-#include <opm/simulators/flow/NewTranFluxModule.hpp>
-#include <opm/simulators/flow/OutputBlackoilModule.hpp>
-#include <opm/simulators/flow/VtkTracerModule.hpp>
 
 #if HAVE_DAMARIS
 #include <opm/simulators/flow/DamarisWriter.hpp>
@@ -51,10 +46,6 @@
 
 #include <tuple>
 
-namespace Opm {
-template <class TypeTag>
-class FlowProblem;
-}
 
 namespace Opm::Properties {
 
@@ -90,15 +81,6 @@ struct EnableThermalFluxBoundaries { using type = UndefinedProperty; };
 template<class TypeTag, class MyTypeTag>
 struct WellModel { using type = UndefinedProperty; };
 
-// Set the problem property
-template<class TypeTag>
-struct Problem<TypeTag, TTag::FlowBaseProblem>
-{ using type = FlowProblem<TypeTag>; };
-
-template<class TypeTag>
-struct Model<TypeTag, TTag::FlowBaseProblem>
-{ using type = FIBlackOilModel<TypeTag>; };
-
 // Select the element centered finite volume method as spatial discretization
 template<class TypeTag>
 struct SpatialDiscretizationSplice<TypeTag, TTag::FlowBaseProblem>
@@ -123,25 +105,6 @@ struct DiscreteFunction<TypeTag, TTag::FlowBaseProblem>
 template<class TypeTag>
 struct GridView<TypeTag, TTag::FlowBaseProblem>
 { using type = typename GetPropType<TypeTag, Properties::Grid>::LeafGridView; };
-
-// Set the material law for fluid fluxes
-template<class TypeTag>
-struct MaterialLaw<TypeTag, TTag::FlowBaseProblem>
-{
-private:
-    using Scalar = GetPropType<TypeTag, Properties::Scalar>;
-    using FluidSystem = GetPropType<TypeTag, Properties::FluidSystem>;
-
-    using Traits = ThreePhaseMaterialTraits<Scalar,
-                                            /*wettingPhaseIdx=*/FluidSystem::waterPhaseIdx,
-                                            /*nonWettingPhaseIdx=*/FluidSystem::oilPhaseIdx,
-                                            /*gasPhaseIdx=*/FluidSystem::gasPhaseIdx>;
-
-public:
-    using EclMaterialLawManager = ::Opm::EclMaterialLawManager<Traits>;
-
-    using type = typename EclMaterialLawManager::MaterialLaw;
-};
 
 // Set the material law for energy storage in rock
 template<class TypeTag>
@@ -213,16 +176,6 @@ template<class TypeTag>
 struct EnableApiTracking<TypeTag, TTag::FlowBaseProblem>
 { static constexpr bool value = false; };
 
-// Use the "velocity module" which uses the Eclipse "NEWTRAN" transmissibilities
-template<class TypeTag>
-struct FluxModule<TypeTag, TTag::FlowBaseProblem>
-{ using type = NewTranFluxModule<TypeTag>; };
-
-// Use the dummy gradient calculator in order not to do unnecessary work.
-template<class TypeTag>
-struct GradientCalculator<TypeTag, TTag::FlowBaseProblem>
-{ using type = DummyGradientCalculator<TypeTag>; };
-
 // store temperature (but do not conserve energy, as long as EnableEnergy is false)
 template<class TypeTag>
 struct EnableTemperature<TypeTag, TTag::FlowBaseProblem>
@@ -276,4 +229,4 @@ struct EnableDebuggingChecks<TypeTag, TTag::FlowBaseProblem>
 
 } // namespace Opm::Properties
 
-#endif // OPM_FLOW_PROBLEM_PROPERTIES_HPP
+#endif // OPM_BASE_FLOW_PROBLEM_PROPERTIES_HPP
